+++ src/models/ApiUser.php
@@ -27,6 +27,7 @@ use luya\admin\aws\UserHistorySummaryActiveWindow;
  * @property integer $api_rate_limit
  * @property string $api_allowed_ips
  * @property integer $api_last_activity
+ * @property boolean $is_request_logger_enabled
  *
  * @author Basil Suter <basil@nadar.io>
  * @since 1.1.0
@@ -93,6 +94,7 @@ final class ApiUser extends User
             [['firstname', 'lastname', 'email'], 'required'],
             [['email'], 'email'],
             [['email', 'auth_token'], 'unique'],
+            [['is_request_logger_enabled'], 'integer'],
         ];
     }
     
@@ -105,6 +107,7 @@ final class ApiUser extends User
             'firstname' => 'text',
             'email' => 'text',
             'lastname' => 'text',
+            'is_request_logger_enabled' => 'toggleStatus',
         ];
     }
     
@@ -114,7 +117,7 @@ final class ApiUser extends User
     public function ngRestScopes()
     {
         return [
-            [['list', 'update', 'create'], ['firstname', 'lastname', 'email']],
+            [['list', 'update', 'create'], ['firstname', 'lastname', 'email', 'is_request_logger_enabled']],
             [['delete'], true],
         ];
     }